Undo moving spltBB
