+++ llvm/lib/Transforms/Utils/BasicBlockUtils.cpp
@@ -1548,81 +1548,6 @@ void llvm::SplitLandingPadPredecessors(BasicBlock *OrigBB,
                                          PreserveLCSSA);
 }
 
-void llvm::spliceBB(IRBuilderBase::InsertPoint IP, BasicBlock *New,
-                    bool CreateBranch, DebugLoc DL) {
-  assert(New->getFirstInsertionPt() == New->begin() &&
-         "Target BB must not have PHI nodes");
-
-  // Move instructions to new block.
-  BasicBlock *Old = IP.getBlock();
-  New->splice(New->begin(), Old, IP.getPoint(), Old->end());
-
-  if (CreateBranch) {
-    auto *NewBr = BranchInst::Create(New, Old);
-    NewBr->setDebugLoc(DL);
-  }
-}
-
-void llvm::spliceBB(IRBuilder<> &Builder, BasicBlock *New, bool CreateBranch) {
-  DebugLoc DebugLoc = Builder.getCurrentDebugLocation();
-  BasicBlock *Old = Builder.GetInsertBlock();
-
-  spliceBB(Builder.saveIP(), New, CreateBranch, DebugLoc);
-  if (CreateBranch)
-    Builder.SetInsertPoint(Old->getTerminator());
-  else
-    Builder.SetInsertPoint(Old);
-
-  // SetInsertPoint also updates the Builder's debug location, but we want to
-  // keep the one the Builder was configured to use.
-  Builder.SetCurrentDebugLocation(DebugLoc);
-}
-
-BasicBlock *llvm::splitBB(IRBuilderBase::InsertPoint IP, bool CreateBranch,
-                          DebugLoc DL, llvm::Twine Name) {
-  BasicBlock *Old = IP.getBlock();
-  BasicBlock *New = BasicBlock::Create(
-      Old->getContext(), Name.isTriviallyEmpty() ? Old->getName() : Name,
-      Old->getParent(), Old->getNextNode());
-  spliceBB(IP, New, CreateBranch, DL);
-  New->replaceSuccessorsPhiUsesWith(Old, New);
-  return New;
-}
-
-BasicBlock *llvm::splitBB(IRBuilderBase &Builder, bool CreateBranch,
-                          llvm::Twine Name) {
-  DebugLoc DebugLoc = Builder.getCurrentDebugLocation();
-  BasicBlock *New = splitBB(Builder.saveIP(), CreateBranch, DebugLoc, Name);
-  if (CreateBranch)
-    Builder.SetInsertPoint(Builder.GetInsertBlock()->getTerminator());
-  else
-    Builder.SetInsertPoint(Builder.GetInsertBlock());
-  // SetInsertPoint also updates the Builder's debug location, but we want to
-  // keep the one the Builder was configured to use.
-  Builder.SetCurrentDebugLocation(DebugLoc);
-  return New;
-}
-
-BasicBlock *llvm::splitBB(IRBuilder<> &Builder, bool CreateBranch,
-                          llvm::Twine Name) {
-  DebugLoc DebugLoc = Builder.getCurrentDebugLocation();
-  BasicBlock *New = splitBB(Builder.saveIP(), CreateBranch, DebugLoc, Name);
-  if (CreateBranch)
-    Builder.SetInsertPoint(Builder.GetInsertBlock()->getTerminator());
-  else
-    Builder.SetInsertPoint(Builder.GetInsertBlock());
-  // SetInsertPoint also updates the Builder's debug location, but we want to
-  // keep the one the Builder was configured to use.
-  Builder.SetCurrentDebugLocation(DebugLoc);
-  return New;
-}
-
-BasicBlock *llvm::splitBBWithSuffix(IRBuilderBase &Builder, bool CreateBranch,
-                                    llvm::Twine Suffix) {
-  BasicBlock *Old = Builder.GetInsertBlock();
-  return splitBB(Builder, CreateBranch, Old->getName() + Suffix);
-}
-
 ReturnInst *llvm::FoldReturnIntoUncondBranch(ReturnInst *RI, BasicBlock *BB,
                                              BasicBlock *Pred,
                                              DomTreeUpdater *DTU) {
